# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BlogSmith",
    "description": "Agentic blogging service \u2014 a 9-stage LangGraph pipeline (Discovery \u2192 Research \u2192 Outline \u2192 Draft \u2192 Critique \u2192 human email gate \u2192 Finalize \u2192 Visuals \u2192 Distribute) that writes human-like, SEO-strong posts with Gemini image generation.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tools/discover": {
      "post": {
        "tags": [
          "tools",
          "tools"
        ],
        "summary": "Discover",
        "operationId": "discover_tools_discover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Discover Tools Discover Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/tools/draft": {
      "post": {
        "tags": [
          "tools",
          "tools"
        ],
        "summary": "Draft Preview",
        "operationId": "draft_preview_tools_draft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Draft Preview Tools Draft Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/tools/preview-image": {
      "post": {
        "tags": [
          "tools",
          "tools"
        ],
        "summary": "Preview Image",
        "operationId": "preview_image_tools_preview_image_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Preview Image Tools Preview Image Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiscoverRequest": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          }
        },
        "type": "object",
        "required": [
          "site_id"
        ],
        "title": "DiscoverRequest"
      },
...
      "DraftRequest": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "keyword": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword"
          },
          "expert_insights": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expert Insights"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "topic"
        ],
        "title": "DraftRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "ImageRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}